Fix receiver connection
